removed external IProps
