--- src/DateTimeRangePicker/DatePicker/Month/IProps.ts
@@ -1,15 +1,0 @@
-import {Moment} from 'moment'
-
-interface IProps {
-    month: number,
-    year: number,
-    fromDate?: Moment,
-    untilDate?: Moment,
-    hoverDate?: Moment,
-    showDaysPreviousMonth: boolean,
-    showDaysNextMonth: boolean,
-    onDaySelected: (date: Moment) => void,
-    onDayHover: (date?: Moment | undefined) => void
-}
-
-export default IProps
