--- src/DateTimeRangePicker/TimePicker/IProps.ts
@@ -1,9 +1,0 @@
-import {Moment} from 'moment'
-
-interface IProps {
-    time?: Moment,
-    step: number,
-    onTimeChanged: (time: Moment) => void
-}
-
-export default IProps
